+++ src/main/java/kr/co/webee/presentation/interestmarket/api/InterestMarketApi.java
@@ -1,4 +1,4 @@
-package kr.co.webee.presentation.market.api;
+package kr.co.webee.presentation.interestmarket.api;
 
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.ArraySchema;
@@ -10,9 +10,9 @@ import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.validation.Valid;
 import kr.co.webee.common.error.ErrorType;
-import kr.co.webee.presentation.market.dto.request.InterestMarketRegisterRequest;
-import kr.co.webee.presentation.market.dto.response.InterestMarketRegisterResponse;
-import kr.co.webee.presentation.market.dto.response.InterestMarketResponse;
+import kr.co.webee.presentation.interestmarket.dto.request.InterestMarketRegisterRequest;
+import kr.co.webee.presentation.interestmarket.dto.response.InterestMarketRegisterResponse;
+import kr.co.webee.presentation.interestmarket.dto.response.InterestMarketResponse;
 import kr.co.webee.presentation.support.annotation.ApiDocsErrorType;
 import kr.co.webee.presentation.support.annotation.UserId;
 import org.springframework.web.bind.annotation.PathVariable;
